+++ components/UIImageLogical/test_ui_image_logical.m
@@ -1,0 +1,12 @@
+[cDirThis, cName, cExt] = fileparts(mfilename('fullpath'));
+
+% Mic
+cDirMic = fullfile(cDirThis, '..', '..');
+addpath(genpath(cDirMic));
+
+purge;
+
+ui = UIImageLogical();
+
+h = figure;
+ui.build(h, 10, 10);